Mocked TaskProgress utils
